feat: unify length results as int64
`len(x)` returned `int64` while every `.len()` member returned `int32`, so
`len(x) == x.len()` was a compile error and ADR-0030's "same operation"
sentence was false. Migrate the maintained length-and-count surface to `int64`
in this batch's breaking window, before the syntax migration.

`String.len()`, `String.byte_len()`, `Vec.len()`, `Map.len()`, and `Set.len()`
now return `int64` through call metadata, semantic checking, analysis and
completion, MIR typing and runtime, direct lowering, and the language server.
The direct backend no longer emits an implicit `int32` range check for these
results; an explicit `as int32` conversion stays checked, which is what the
retained `int32` index domain uses at its boundaries.

Two defects found by review during the work, each reproduced first:

- MIR member dispatch cloned the complete place receiver before reading a
  length. Six clone-count regressions failed with one snapshot apiece for the
  five member forms and source-lowered free `len(...)`. A borrowed place fast
  path now reads every length without cloning or moving the receiver, keeping
  `String.byte_len()`'s O(1) contract.
- Vec/Map/Set completion rows could duplicate where the hardcoded surface and
  the catalog overlapped. Catalog entries are appended only when the member
  name is not already present, and both compiler and LSP tests now require
  unique rows with exact `int64` details.

Count/index audit, as the ticket requires: already `int64` were free `len(...)`
and `enumerate` positions; migrated here are the five public member lengths;
deliberately retained as `int32` are Range endpoints and yields, Vec index
inputs and internal cursors, and `WaitAny`/`WaitAll` indices, which form one
coordinated index domain reached by explicit checked casts, plus `bytes.Error`
offsets and JSON error line/column/depth payloads, which are bounded diagnostic
ABIs. Also retained: supervisor restart counts and configuration, Queue
capacity, bounded network and process byte-count inputs, HTTP status, process
exit and signal codes, and `main` exit codes — codes are not collection counts,
and the bounded operational parameters need their own coordinated API decision.

Resource ceilings were not widened. The codec path now names
`MAX_CODEC_OUTPUT_LEN`/`checked_codec_output_len` and secure randomness names
`RequestExceedsCeiling`, so the renamed run-fail fixtures pin a request ceiling
rather than describing it as a maximum representable `Vec` length.

ADR-0030 is Accepted with the length-unification amendment; ADR-0004, ADR-0020,
and ADR-0023 record the affected contracts. The Manual, Learn, READMEs,
examples, and tutorials teach free/member type-and-value equality, Unicode
scalar versus UTF-8 byte counts, and explicit checked narrowing, and the
reference gate freezes those contracts while rejecting stale `int32` length
prose.

The exact full-repository `npm run ci` gate is green, with compiler coverage at
64,612/67,239 lines, 4,179/4,315 functions, and 94,761/100,470 regions, above
the frozen 96.07/96.81/94.29 floors, which are not raised during the migration.

+++ crates/aurora-compiler/src/mir.rs
@@ -6529,7 +6529,7 @@ impl<'a> Lowerer<'a> {
             return Some(Type::named("String"));
         }
         match (name.as_str(), field) {
-            ("String", "len" | "byte_len") => Some(Type::named("int32")),
+            ("String", "len" | "byte_len") => Some(Type::named("int64")),
             ("String", "contains") | ("String", "starts_with") | ("String", "ends_with") => {
                 Some(Type::named("bool"))
             }
@@ -6546,7 +6546,7 @@ impl<'a> Lowerer<'a> {
                 "Option".to_string(),
                 vec![Type::named("String")],
             )),
-            ("Vec", "len") => Some(Type::named("int32")),
+            ("Vec", "len") => Some(Type::named("int64")),
             ("Vec", "is_empty") => Some(Type::named("bool")),
             ("Vec", "clone") => Some(Type::Named("Vec".to_string(), args.clone())),
             ("Vec", "push") | ("Vec", "extend") | ("Vec", "clear") | ("Vec", "reverse") => {
@@ -6562,13 +6562,13 @@ impl<'a> Lowerer<'a> {
                         .unwrap_or_else(|| Type::named("Unknown"))],
                 ))
             }
-            ("Set", "len") => Some(Type::named("int32")),
+            ("Set", "len") => Some(Type::named("int64")),
             ("Set", "is_empty") => Some(Type::named("bool")),
             ("Set", "clone") => Some(Type::Named("Set".to_string(), args.clone())),
             ("Set", "contains") | ("Set", "insert") | ("Set", "remove") => {
                 Some(Type::named("bool"))
             }
-            ("Map", "len") => Some(Type::named("int32")),
+            ("Map", "len") => Some(Type::named("int64")),
             ("Map", "is_empty") => Some(Type::named("bool")),
             ("Map", "clone") => Some(Type::Named("Map".to_string(), args.clone())),
             ("Map", "contains_key") => Some(Type::named("bool")),
